pyproject.toml
[tool.poetry]
name = "moviescrapper"
version = "0.1.0"
description = ""
authors = ["Esperanza Dueñas <esperanzaduenas008@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.4.0"
webdriver-manager = "^3.8.3"
numpy = "^1.23.2"
pandas = "^1.4.3"
Flask = "^2.2.2"
mariadb = "^1.1.4"
Flask-MySQLdb = "^1.0.1"
art = "^5.7"
rich = "^12.5.1"
text2art = "^0.2.0"
schedule = "^1.1.0"
openpyxl = "^3.0.10"
xlrd = "^2.0.1"
fuzzywuzzy = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
